Print IIDs
Apparently AtlDumpIID has been removed from the Win32 API since 2003. Not shocking since it was undocumented; MS never intended us to use it in the first place. Weeell my predecessors did, and I would kind of like to see that information again myself. So how fortunate it is that Windows XP's source code was leaked! I just plucked the source for the function straight out and pasted it in. Now we have a nice polyfill for AtlDumpIID and I can log these interface names again.