Proxy along EnumObjects if not our requested path
Found out Windows will call EnumObjects on the parent of the path we asked for before calling EnumObjects on the real thing. That first call has to act normal! Can't enumerate ADSes then. Unfortunately the only way I can think of to tell them apart though is through heuristics on the flags Windows passes me. Hope this never breaks

Being real aggressive about the sets of flags I take in so that the program yells whenever it sees a set of flags I haven't seen before so I can learn from it